+++ starsim/sim.py
@@ -49,6 +49,7 @@ class Sim(sc.prettyobj):
 
         # Placeholders for plug-ins: demographics, diseases, connectors, analyzers, and interventions
         # Products are not here because they are stored within interventions
+        if demographics == True: demographics = [ss.Births(), ss.Deaths()]  # Use default assumptions for demographics
         self.demographics = ss.ndict(demographics, type=ss.BaseDemographics)
         self.diseases = ss.ndict(diseases, type=ss.Disease)
         self.networks = ss.ndict(networks, type=ss.Network)
@@ -265,28 +266,34 @@ class Sim(sc.prettyobj):
 
         # Figure out if it's in the sim pars or provided directly
         attr_plugins = getattr(self, plugin_name)  # Get any plugins that have been provided directly
-        if attr_plugins is None or len(attr_plugins) == 0:  # None have been provided directly
 
-            # See if they've been provided in the pars dict
-            if self.pars.get(plugin_name):
+        # See if they've been provided in the pars dict
+        if self.pars.get(plugin_name):
 
-                par_plug = self.pars[plugin_name]
+            par_plug = self.pars[plugin_name]
 
-                # String: convert to ndict
-                if isinstance(par_plug, str):
-                    plugins = ss.ndict(dict(name=par_plug))
+            # String: convert to ndict
+            if isinstance(par_plug, str):
+                plugins = ss.ndict(dict(name=par_plug))
 
-                # List or dict: convert to ndict
-                elif sc.isiterable(par_plug) and len(par_plug):
-                    if isinstance(par_plug, dict) and 'type' in par_plug and 'name' not in par_plug:
-                        par_plug['name'] = par_plug['type'] # TODO: simplify/remove this
-                    plugins = ss.ndict(par_plug)
+            # List or dict: convert to ndict
+            elif sc.isiterable(par_plug) and len(par_plug):
+                if isinstance(par_plug, dict) and 'type' in par_plug and 'name' not in par_plug:
+                    par_plug['name'] = par_plug['type'] # TODO: simplify/remove this
+                plugins = ss.ndict(par_plug)
 
-            else:  # Not provided directly or in pars
-                plugins = {}
-        else:
-            plugins = attr_plugins
+        else:  # Not provided directly or in pars
+            plugins = {}
 
+        # Check that we don't have two copies
+        for attr_key in attr_plugins.keys():
+            if plugins.get(attr_key):
+                errormsg = f'Sim was created with {attr_key} module, cannot create another through the pars dict.'
+                raise ValueError(errormsg)
+
+        plugins = sc.mergedicts(plugins, attr_plugins)
+
+        # Process
         processed_plugins = sc.autolist()
         for plugin in plugins.values():
 
@@ -332,6 +339,22 @@ class Sim(sc.prettyobj):
         for dem_mod in demographics:
             dem_mod.initialize(self)
             self.results[dem_mod.name] = dem_mod.results
+
+        # Count how many of each kind of demographic module we have
+        demdict = {'births': ss.Births, 'pregnancy': ss.Pregnancy, 'deaths': ss.Deaths}
+        mod_names = dict()
+        for demname, demtype in demdict.items():
+            mod_names[demname] = [d.name for d in demographics if isinstance(d, demtype)]
+
+            # Validation
+            if len(mod_names[demname]) > 1:
+                if len(mod_names[demname]) == len(set(mod_names[demname])):  # No duplicate names, raise warning
+                    ss.warn(f'Two instances of {demname} module added to the sim; was this intentional?')
+                else:
+                    errormsg = (f'Cannot add two identically-named {demname} modules to a sim.\n '
+                                f'Demographic modules are: \n{sc.newlinejoin(mod_names[demname])}.\n'
+                                f'Tip: if using demographic modules, do not use birth and death rates in the sim pars.')
+                    raise ValueError(errormsg)
 
         # Ensure they're stored at the sim level
         self.demographics = ss.ndict(*demographics)
@@ -454,8 +477,8 @@ class Sim(sc.prettyobj):
             self.people.remove_dead(self)
 
         # Update demographic modules (create new agents from births/immigration, schedule non-disease deaths and emigration)
-        for module in self.demographics.values():
-            module.update(self)
+        for dem_mod in self.demographics.values():
+            dem_mod.update(self)
 
         # Carry out autonomous state changes in the disease modules. This allows autonomous state changes/initializations
         # to be applied to newly created agents